[feat]: Refactored native duration type to fix deserializaiton problems of json's coming from tendermint
