+++ rpc/tests/kvstore_fixtures/incoming/consensus_params.json
@@ -20,36 +20,15 @@
         ]
       },
       "synchrony": {
-        "message_delay": {
-          "secs": 0,
-          "nanos": 505000000
-        },
-        "precision": {
-          "secs": 12,
-          "nanos": 0
-        }
+        "message_delay": "505000000",
+        "precision": "12000000000"
       },
       "timeout": {
-        "propose": {
-          "secs": 3,
-          "nanos": 0
-        },
-        "propose_delta": {
-          "secs": 0,
-          "nanos": 500000000
-        },
-        "vote": {
-          "secs": 1,
-          "nanos": 0
-        },
-        "vote_delta": {
-          "secs": 0,
-          "nanos": 500000000
-        },
-        "commit": {
-          "secs": 1,
-          "nanos": 0
-        },
+        "propose": "3000000000",
+        "propose_delta": "500000000",
+        "vote": "1000000000",
+        "vote_delta": "500000000",
+        "commit": "1000000000",
         "bypass_commit_timeout": false
       },
       "version": {}
