+++ testgen/src/consensus.rs
@@ -1,6 +1,5 @@
-use std::time::Duration;
 use tendermint::consensus::params::{SynchronyParams, TimeoutParams, VersionParams};
-use tendermint::{block, consensus, evidence, public_key::Algorithm};
+use tendermint::{block, consensus, duration::Duration, evidence, public_key::Algorithm};
 
 /// Default consensus params modeled after Go code; but it's not clear how to go to a valid hash
 /// from here
@@ -12,7 +11,7 @@ pub fn default_consensus_params() -> consensus::Params {
         },
         evidence: evidence::Params {
             max_age_num_blocks: 100000,
-            max_age_duration: evidence::Duration(std::time::Duration::new(48 * 3600, 0)),
+            max_age_duration: Duration(std::time::Duration::new(48 * 3600, 0)),
             max_bytes: 1048576,
         },
         validator: consensus::params::ValidatorParams {
